+++ execFileSystem.go
@@ -5,16 +5,19 @@ import (
 	"fmt"
 	"html/template"
 	"net/http"
-	"text/template/parse"
 )
 
 // FileSystemExecutor loads view templates as a path from a Go HTML template file.
 // The underlying file system is abstracted through the http.FileSystem interface to allow for
 // in-memory use.
+//
+// The optional KeyedString map will be checked before the loader attempts to use the FS
+// instance when obtain a template string
 type FileSystemExecutor struct {
-	FS    http.FileSystem
-	Funcs template.FuncMap
-	exec  Executor
+	FS          http.FileSystem
+	Funcs       template.FuncMap
+	KeyedString map[string]string
+	exec        Executor
 }
 
 // NewViewHandler creates a ViewHandler from a View endpoint definition treating
@@ -37,7 +40,6 @@ func (fe *FileSystemExecutor) constructTemplate(view *View) (*template.Template,
 		return nil, nil
 	}
 	var out *template.Template
-	cache := make(map[string]*parse.Tree)
 	buffer := new(bytes.Buffer)
 
 	queue := viewQueue{}
@@ -45,6 +47,11 @@ func (fe *FileSystemExecutor) constructTemplate(view *View) (*template.Template,
 
 	for !queue.empty() {
 		v, _ := queue.next()
+		for _, sub := range v.SubViews {
+			if sub != nil {
+				queue.add(sub)
+			}
+		}
 		var t *template.Template
 		if out == nil {
 			out = template.New(v.Defines).Funcs(fe.Funcs)
@@ -53,9 +60,14 @@ func (fe *FileSystemExecutor) constructTemplate(view *View) (*template.Template,
 			t = out.New(v.Defines)
 		}
 
-		if tree, ok := cache[v.Template]; ok {
-			t.AddParseTree(v.Defines, tree)
-		} else {
+		var (
+			templateString string
+			foundKey       bool
+		)
+		if fe.KeyedString != nil {
+			templateString, foundKey = fe.KeyedString[v.Template]
+		}
+		if !foundKey {
 			buffer.Reset()
 			file, err := fe.FS.Open(v.Template)
 			if err != nil {
@@ -64,25 +76,22 @@ func (fe *FileSystemExecutor) constructTemplate(view *View) (*template.Template,
 					v.Template, err.Error(),
 				)
 			}
-			_, err = buffer.ReadFrom(file)
+
+			templateString, err = readStringAndClose(buffer, file)
 			if err != nil {
 				return nil, fmt.Errorf(
 					"Failed to read contents of template file '%s', error %s",
 					v.Template, err.Error(),
 				)
 			}
-			_, err = t.Parse(buffer.String())
-			if err != nil {
-				return nil, fmt.Errorf(
-					"Failed to parse contents of template file '%s', error %s",
-					v.Template, err.Error(),
-				)
-			}
 		}
-		for _, sub := range v.SubViews {
-			if sub != nil {
-				queue.add(sub)
-			}
+
+		_, err := t.Parse(templateString)
+		if err != nil {
+			return nil, fmt.Errorf(
+				"Failed to parse contents of template file '%s', error %s",
+				v.Template, err.Error(),
+			)
 		}
 	}
 	return out, nil