+++ execFiles_test.go
@@ -2,6 +2,7 @@ package treetop
 
 import (
 	"html/template"
+	"net/http"
 	"net/http/httptest"
 	"strings"
 	"testing"
@@ -104,17 +105,17 @@ func TestFileExecutor_NewViewHandler(t *testing.T) {
 			},
 		},
 		{
-			name: "malformed file",
+			name: "error template file",
 			getHandler: func(exec ViewExecutor) ViewHandler {
-				return exec.NewViewHandler(NewView("testdata/malformed.html", Noop))
+				return exec.NewViewHandler(NewView("testdata/missingFunc.html", Noop))
 			},
 			expectPage:     "Not Acceptable\n",
 			expectTemplate: "Not Acceptable\n",
 			expectErrors: []string{
-				`Failed to parse contents of template file 'testdata/malformed.html', ` +
-					`error template: :2: function "functhatdoesnexist" not defined`,
-				`Failed to parse contents of template file 'testdata/malformed.html', ` +
-					`error template: :2: function "functhatdoesnexist" not defined`,
+				`Failed to parse contents of template file 'testdata/missingFunc.html', ` +
+					`error template: :2: function "func_that_does_not_exist" not defined`,
+				`Failed to parse contents of template file 'testdata/missingFunc.html', ` +
+					`error template: :2: function "func_that_does_not_exist" not defined`,
 			},
 		},
 		{
@@ -231,5 +232,54 @@ func TestFileExecutor_FuncMap(t *testing.T) {
 	</div>
 	`)) {
 		t.Errorf("Expecting title case, got\n%s", gotPage)
+	}
+}
+
+func TestFileExecutor_KeyedString(t *testing.T) {
+	exec := FileExecutor{
+		KeyedString: map[string]string{
+			"local://titles.html": `<h1>{{ . }}</h1>`,
+		},
+	}
+	v := NewView("local://titles.html", Constant("Test Title"))
+	h := exec.NewViewHandler(v)
+	rec := httptest.NewRecorder()
+	h.ServeHTTP(rec, mockRequest("/some/path", "*/*"))
+	gotPage := stripIndent(sDumpBody(rec))
+	if gotPage != stripIndent(strings.TrimSpace("<h1>Test Title</h1>")) {
+		t.Errorf("Expecting title case, got\n%s", gotPage)
+	}
+}
+
+func TestFileExecutor_UsingSameTemplate(t *testing.T) {
+	exec := FileExecutor{
+		KeyedString: map[string]string{
+			"local://common.html": `<h1>Common {{ . }}</h1>`,
+		},
+	}
+	v := NewView("./testdata/base.html", func(rsp Response, req *http.Request) interface{} {
+		return struct {
+			Content interface{}
+			PS      interface{}
+		}{
+			Content: rsp.HandleSubView("content", req),
+			PS:      rsp.HandleSubView("ps", req),
+		}
+	})
+	v.NewDefaultSubView("content", "local://common.html", Constant("Content Data"))
+	v.NewDefaultSubView("ps", "local://common.html", Constant("Postscript Data"))
+	h := exec.NewViewHandler(v)
+	rec := httptest.NewRecorder()
+	h.ServeHTTP(rec, mockRequest("/some/path", "*/*"))
+	gotPage := strings.TrimSpace(stripIndent(sDumpBody(rec)))
+	expectPage := strings.TrimSpace(stripIndent(`
+	<html><body>
+	<h1>Common Content Data</h1>
+
+	<h1>Common Postscript Data</h1>
+	</body></html>
+	`))
+	if gotPage != expectPage {
+		t.Errorf("Expecting\n%s\ngot\n%s", expectPage, gotPage)
 	}
 }