+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+## [0.3.1] - 2020-04-07
+
+### New Features
+
+- `FileExecutor` and `FileSystemExecutor` now have an optional `KeyedString` property to mix files with hard coded templates
+- Added `Template` interface so that `TemplateHandler` no longer depends directly on html/template
+
+### Bugfix
+
+- Template cache not used/not working when constructing template in file executors
+- Files not being closed properly in FS executor
+
+
 ## [0.3.0] - 2020-03-10
 
 Protocol and API overhaul, improve docs and examples, transitioning from prototype